refactor(protocol): extract PendingDecryptionQueue struct and fix duplicate import
Encapsulate the 8 pending-queue fields from OfflineProtocol into a
dedicated PendingDecryptionQueue struct with self-contained methods,
reducing coupling and improving cohesion. Also remove duplicate
std::time::Duration import in tests.
